+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 - Added studio-viewPublishedRSS-[PodcastUID] permission to check accessing to RSS
 - Added checkAccessToVolumes general setting which checks saveAssets:[VolumeUID] permission before listing volumes on asset index page
 - Added lastBuildDate for podcast RSS
-- Added support for content:encoded
+- Added support importing from content:encoded
+- Added support importing from itunes:keywords
 - Now each podcast has two settings pages (general settings, episode settings)
 - Improved fetch from RSS via CURL 
 - Improved fetch from RSS progress notifications